mac80211: don't overwrite QoS TID of injected frames
Currently ieee80211_set_qos_hdr overwrites the QoS TID of injected
frames based on the value assigned to skb->priority. The commit
753ffad3d624 ("mac80211: fix TID field in monitor mode transmit")
worked around this problem by setting skb->priority to the TID of
the injected frame, which assured the QoS TID will not be changed
to a different value. Unfortunately, this workaround complicates
the handling of injected frames because we can't set skb->priority
without affecting the TID value of injected frames.

To avoid this, and to simplify the next patch, detect if a frame is
injected in ieee80211_set_qos_hdr and if so do not change its QoS
field.

Signed-off-by: Mathy Vanhoef <Mathy.Vanhoef@kuleuven.be>