Experimental support for kernel 5.13
The following changes will make the drivers compile on kernel 5.13
which is currently being used by Ubuntu 20.04. However, the
injection tests were not run, so it's unclear how well these drivers
will actually work. Hence support is only experimental and kernels
below or equal to 5.12 are still recommended for the test tool.